+++ app/src/main/java/com/example/workout_tracker/data/dao/RoutineDAO.kt
@@ -4,10 +4,8 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import androidx.room.Transaction
 import androidx.room.Update
 import com.example.workout_tracker.data.model.Routine
-import com.example.workout_tracker.data.model.RoutineWithWorkouts
 
 @Dao
 interface RoutineDAO {
@@ -21,7 +19,7 @@ interface RoutineDAO {
     fun delete(routine: Routine)
 
     @Query("SELECT * FROM Routine WHERE id = :id")
-    fun getById(id: Int): Routine
+    fun getById(id: Long): Routine
 
     @Query("DELETE FROM Routine")
     fun deleteAll(): Int
