+++ app/src/main/java/com/example/workout_tracker/data/dao/WorkoutDAO.kt
@@ -4,11 +4,8 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import androidx.room.Transaction
 import androidx.room.Update
-import com.example.workout_tracker.data.model.RoutineWithWorkouts
 import com.example.workout_tracker.data.model.Workout
-import com.example.workout_tracker.data.model.WorkoutWithExercises
 
 @Dao
 interface WorkoutDAO {
@@ -22,7 +19,7 @@ interface WorkoutDAO {
     fun delete(workout: Workout)
 
     @Query("SELECT * FROM Workout WHERE id = :id")
-    fun getById(id: Int): Workout
+    fun getById(id: Long): Workout
 
     @Query("DELETE FROM Workout")
     fun deleteAll(): Int
